Moved constraints to its own package
